+++ guides/Included Pieces/IncludedEvents.md
@@ -14,13 +14,13 @@ Replies with the reason why the command was inhibited.
 
 [events/commandInhibited.js](https://github.com/dirigeants/klasa/blob/master/src/events/commandInhibited.js)
 
-## settingsUpdateEntry
+## settingsUpdate
 
 Synchronises the user settings between all shards, if the bot is sharded.
 
 **Source:**
 
-[events/settingsUpdateEntry.js](https://github.com/dirigeants/klasa/blob/master/src/events/settingsUpdateEntry.js)
+[events/coreSettingsUpdate.js](https://github.com/dirigeants/klasa/blob/master/src/events/coreSettingsUpdate.js)
 
 ## debug
