+++ src/lib/settings/gateway/Gateway.js
@@ -114,6 +114,7 @@ class Gateway extends GatewayStorage {
 				if (cache) {
 					cache.existenceStatus = true;
 					cache._patch(entry);
+					this.client.emit('settingsSync', cache);
 				}
 			}
